requirements.txt
requests>=2.28.0
PyPDF2>=3.0.0
pdfplumber==0.11.4
pandas>=1.5.0
numpy>=1.23.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
google-generativeai>=0.5.0
nltk>=3.8.0
spacy>=3.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
tabulate==0.9.0
groq
fastapi